challenges/web-kinjutsu/solution/requirements.txt
PyJWT==2.8.0
colorama==0.4.6
